Split dashboard UI settings into focused tabs
