+++ frontend/src/views/Settings.vue
@@ -263,12 +263,30 @@ const settingsDestinations = [
   },
   {
     id: "dashboard-ui",
-    label: "UI, theme, clock, and notifications",
-    path: "Dashboard / UI, Theme & Clock",
+    label: "Appearance and kiosk UI",
+    path: "Dashboard / Appearance",
     category: "dashboard",
     tabKey: "settings_tab_dashboard",
-    tab: "ui",
-    keywords: ["ui", "theme", "clock", "kiosk", "notifications", "weather"],
+    tab: "appearance",
+    keywords: ["ui", "theme", "appearance", "kiosk", "headers", "dark mode"],
+  },
+  {
+    id: "dashboard-clock",
+    label: "Clock and status bar",
+    path: "Dashboard / Clock",
+    category: "dashboard",
+    tabKey: "settings_tab_dashboard",
+    tab: "clock",
+    keywords: ["clock", "status bar", "date", "seconds", "weather"],
+  },
+  {
+    id: "dashboard-notifications",
+    label: "Notifications",
+    path: "Dashboard / Notifications",
+    category: "dashboard",
+    tabKey: "settings_tab_dashboard",
+    tab: "notifications",
+    keywords: ["notifications", "keyboard feedback", "mode indicator"],
   },
   {
     id: "content-calendars",
